requirements.txt
requests>=2.31.0
beautifulsoup4
flask
schedule
watchdog
python-crontab
python-dotenv>=0.21.0
deepdiff>=6.3.0 